Update to match latest question types and add create api call to creator
