+++ frontend/src/components/creator-modal/multiple-choice/multiple-choice.scss
